pkg/utils/types/pointer_test.go
/*
 * Tencent is pleased to support the open source community by making TKEStack available.
 *
 * Copyright (C) 2012-2019 Tencent. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
 * this file except in compliance with the License. You may obtain a copy of the
 * License at
 *
 * https://opensource.org/licenses/Apache-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations under the License.
 */

package types

import (
	"testing"

	"github.com/stretchr/testify/require"
)

func TestPtr(t *testing.T) {
	i32 := Int32Ptr(1)
	require.Equal(t, int32(1), *i32)

	i64 := Int64Ptr(1)
	require.Equal(t, int64(1), *i64)

	b := BoolPtr(true)
	require.Equal(t, true, *b)

	s := StringPtr("123")
	require.Equal(t, "123", *s)
}
